requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
supabase==1.0.3
requests==2.26.0
python-jose==3.3.0
Flask-Login==0.5.0
